put files into folders
